fix(sets): simplify imageset intersection handler
The imageset intersection handler was overly complex and made faulty
assumptions about the expression to be inverted. This commit simplifies
the intersect handler so that it is limited to inverting linear
transformations e.g.:

   {a*x + b: x in S} n [c, d]

is now transformed to:

   {a*x + b: x in (S n [(c - b)/a, (d - b)/a])}

So if the base set S can compute an intersection with an interval, then
the imageset intersection can be expressed in terms of that.