quantum: Added logic in qapply to handle Sum.
Previously, quantum expressions with Sums were not being handled. This change introduces full handling for Sums, both at the top level and in the qapply_Mul section. The implementation brings the sums to the left of products and calls qapply_Mul on the function of the Sum. To keep backwards compatibiilty, I have introduced a new sum_doit keyword argument to qapply that defaults to False. When set to True, this calls Sum.doit() on any Sums found in the process of applying operators to states.

Note: qapply does need some refactoring. After these changes I have some ideas about that, but I don't want to mix those improvements with this fix.