quantum: Improve tensor product handling
Previously this PR tried to combine products of kets or bras into tensor products using normal multiplication. This was leading to a number of confusing situation where regular* multiplication could mean different things in different situations (in quantum expressions). This has been resolved by transforms now raising TypeError if you try to combine bras and kets this way. Instead, users need to explicitly use TensorProduct.

In the future, I would like to introduce a new binary operator that is a shorthand for TensorProduct, but that is work for later.

Another side effect of this work is that the tensor_product_simp functions and their helpers can be deprecated, which this commit also does.