fix(polys): change ring series tests using negative/fractional exponents.
The ring_series module creates polynomials with fractional or negative
exponents. These are not polynomials and it should not be permitted to
create them with PolyElement. Instead a new domain for series with
fractional and negative exponents should be used. This commit adjusts
the ring series tests to avoid negative powers and marks some tests as
XFAIL since they make unavoidable use of fractional powers.