Assuming the number given to `_ecm_one_factor` is composite
Currently, we assume that the number `n` given to
`sympy.ntheory.ecm._ecm_one_factor` is a natural number. If `n` is prime,
it simply returns `n`. However, when called from `factorint`,
it is already known that `n` is not a prime number, making a primality
check within `_ecm_one_factor` unnecessary.

Therefore, we have modified `_ecm_one_factor` to assume that the given
number is composite and to delegate the primality check to `ecm`.
This specification is consistent with other factorization algorithms.

In `ecm`, we also check whether the number is a perfect power.
This is because if a perfect power is passed to `_ecm_one_factor`,
it may fail to find a factor.