physics.control: fixed a bug in rewrite of tf to ss and revised tests
There was a bug in the rewrite of transfer function(tf) to state space (ss)
model as mentioned in issue #26827, where because of a mismatch in tf model
used in the rewrite implementation and the one in documentation it lead to
an incorrect conversion, i.e. mainly because of the assumption in the tf
model used in the implementation that, the leading coefficent of highest
degree term in denominator polynomial was unity, but in the docs it was
generic. So refactored the code to fix this issue by scaling down the
coefficients when constructing the ss matrices (only for C, D and for A it
exists already) and also revised the test in order to check for this bug.

Example:

>>> import sympy
>>> from sympy.physics.control.lti import StateSpace, TransferFunction
>>>
>>> s, K, T1 = sympy.symbols('s, K, T1')
>>>
>>> G = TransferFunction(K, T1*s + 1, s)
>>> print(G)
TransferFunction(K, T1*s + 1, s)
>>>
>>> Gss = G.rewrite(StateSpace)
>>> print(Gss)
StateSpace(Matrix([[-1/T1]]), Matrix([[1]]), Matrix([[K/T1]]), Matrix([[0]]))
>>>
>>> Gss.rewrite(TransferFunction)
[[TransferFunction(K, T1*s + 1, s)]]

Previously, Gss was (which is wrong),

>>> StateSpace(Matrix([[-1/T1]]), Matrix([[1]]), Matrix([[K]]), Matrix([[0]]))