series: Added a function pade_approximant and tests
The pade approximant is similar to a taylor series, although rather than approximating an expression with a polynomial of given degree, a pade approximant approximates an expression with a rational function whose numerator and denominator each have given degree.

Added a pade_approximant function. This function allows users to easily compute the [n/m] pade approximant of a given expression with a similar interface to computing a series expansion for an expression. I wrote this because I used to use the function "PadeApproximant" frequently in mathematica and found it very useful. The code for computing the approximant is adapted from the source code for scipy.interpolate.pade.

Example:

In [1]: import sympy as sp

In [2]: x = sp.symbols('x')

In [3]: pade_approximant(sp.exp(x), x, 0, 2, 1)

Out [3]: (x**2/6 + 2*x/3 + 1)/(1 - x/3)

The command "pade_approximant(sp.exp(x), x, 0, 2, 1)" computes the [2/1] approximant of exp(x) around the point x = 0.