`_check_smoothness` has been modified to return a vector and `num`
In `_check_smoothness`, depending on the value of `num`, it returns either a
pair of a vector and `True`, a pair of `num` and `False`, or a pair of `None`
and `None`. It would be clearer to simply return the vector and `num` and
handle the branching in `_trial_division_stage`. In fact, the current code
already performs branching based on the return value, making it redundant.