Enable complex number support in `eval_zeta_function`
In `eval_zeta_function`, conversion from `Sum` to `zeta` is performed whenever
possible. Currently, the conditions `s > 1` and `q > 0` are set for `s` and
`q`, but these conditions do not handle cases where `s` or `q` are complex
numbers correctly. According to Wikipedia, the Hurwitz zeta function is defined
under the conditions `Re(s) > 1` and `q != 0, -1, -2, ...`. Therefore, we have
modified the conditions accordingly.