Fix Rust printing of sympy.sign
This is half issue, half PR - I'm not 100% sure exactly how this should be correctly implemented, and I'm not necessarily going to have time to test this further, but the fix is small enough and the code change demonstrates the issue better than an issue would.

As I understand it, `sympy.sign` is defined to return 0 for 0.  Other code printers also implement it this way, see for example https://github.com/sympy/sympy/pull/14010.  Rust's `signum` returns 1 or -1 for 0 (depending on the sign of the 0, for floats with signed zeroes), which isn't correct.