Add documentation warning to parse_expr() about internal eval usage
This commit adds a warning the to the parse_expr() function's
documentation to include the same warning admonition that the sympify()
function contains. Internally parse_expr() also uses eval() and is
subject to the same security concerns as sympify, however the
documentation did not make this clear. This commit fixes this
documentation gap so it's clear that you shouldn't use this function
with untrusted input.