Added handling for cases where `floor` and `ceiling` can be evaluated as constants
For example, `floor(1/n)` always evaluates to 0 when `n` is a real number
greater than 1. In such cases, where `floor` or `ceiling` can be determined
to be a small constant under certain assumptions, they are now evaluated
accordingly.
This is equivalent to the logic originally implemented in `Mod.eval`, but it
has been moved here as it can be more generally applied to `floor`. Since
there was a bug in the original implementation that could lead to incorrect
evaluations, this PR does not simply move the code as-is, but instead
rewrites it from scratch.