matrix: Fix is_hermitian for block matrix handling
The function is_hermitian doesn't work for block matrices like [[A]] where A is an HermitianOperator. Before this commit, Matrix[[A]].is_hermitian returned False, whereas Matrix[[A]] is hermitian because A is hermitian.
This commit fixes this issue by checking if M[i, j] - Dagger(M[j, i]) is zero, which is the definition of an hermitian matrix.