Revert passing init_printing settings to the pretty printer
This is a partial revert of PR #26997.

This caused a regression where settings that are not supported by the pretty
printer but that are supported by other printers would lead to an error (there
are many such settings in the LaTeX printer). A more correct fix would require
some deeper refactoring, so for now we are just reverting this change.

The changes from #26997 updating the minimum IPython version have not been
reverted.

That PR did not add any tests for this change and this does not add any tests
either, as testing IPython/Jupyter interactive behavior is not
straightforward.

Fixes #27992