Improve _eval_is_hermitian _eval_is_antihermitian
This commit will improve _eval_is_hermitian and _eval_is_antihermitian.
With this improvement, _eval_is_hermitian will check if an expression is hermitian by
checking the definition of an hermitian object ie A.adjoint() = A.
To do that, we use a pattern matching approach to avoid the creation of
new expressions. We will check if the expression is "Palindrome-like"
with adjoint.