Update column.py
The base of the Column class has been written. A column with a certain length, area and elastic modulus can be initialized. Methods have also been added to apply supports and forces. These methods add the singularity functions to the load equations. The user can get the load equation through the 'load' property.