quality: Add ruff rule to disallow catching Exception
It is almost always incorrect to use bare except or to catch Exception
or BaseException. This commit adds a ruff rule to disallow this. A bare
`except:` is already disallowed by rule E722 but this extends that to
`except Exception` and `except BaseException`.

Some existing cases in the codebase are left as is because they are
exceptional such as handling exceptions in the old test runner. Some
cases are left because it just is not clear what exceptions they are
intended to catch (precisely why except Exception should not be used).