fix(matrices): cancel gcd exactly in Matrix.inv
The Matrix.inv() method converts from e.g. RR[x] to QQ[x] if the domain
is inexact and not purely numeric like RR and CC (that use partial
pivoting instead). This allows computing the inverse exactly using e.g.
fraction-free methods. However the denominator form the fraction-free
method needs to be cancelled with the matrix in the exact domain QQ[x]
before converting back to the inexact domain RR[x]. Otherwise GCD fails
in the inexact domain and the expressions will have higher degree
numerators and denominators that would be difficult to simplify after.