fix(polys): factor the base in roots_binomial
Given a polynomial of the form a*x**n + b = 0 the roots can be found
easily as x = (-b/a)**(1/n) muliplied by roots of unity. This formula
introduces radicals though with symbols which are generally bad.
Factorising the base can potentially reveal that it is an nth power
e.g.: x**n = c**n meaning that the solutions are x = c multiplied by nth
roots of unity without introducing any radicals with needing to put any
symbols under the radical.