typ: fix type of dup_from_dict
Previously dup_from_dict would accept either a dict with int keys or a
dict with tuple-of-int keys. This commit fixes the type to be
tuple-of-int leaving dup_from_raw_dict to handle the int case. The
amiguous types are handled at a higher level in the Poly constructor or
the ANP constructor instead.