fix(polys): remove special zero handling in dup_to_dict
The functions `dup_to_dict`, `dup_to_raw_dict` and `dmp_to_dict` are
changed to no longer have a zero parameter that changes how the zero
polynomial is handled. Places that call these functions are also
changed to pass the domain parameter explicitly although there are
two places in densebasic where it is not clear what the correct
domain should be.