feat(polys): Add abstract RingExtension Domain
A RingExtension is a ring whose elements are polynomials over another
domain. Various operations use this kind of domain generically with the
expectation that the elements have a `.to_dict()` method. This commit
adds the abstract domain class `RingExtension` that has a `K.to_dict(e)`
method that can be used in a well defined way in typed code.

Concrete implementations of `RingExtension` are polynomial rings,
algebraic field extensions and the Gausian integers and rationals.