fix(printing): handle python printing of piecewise with "assign_to"
When printing an assignment where the RHS is a piecewise assignment, the current PythonCodePrinter prints two assignments in the ternary if-else form, which is invalid Python. This commit adds a check in _print_Assignment which tests whether the RHS is a Piecewise expression, and if so, returns "lhs = rhs", letting the Piecewise operator be correctly processed as an expression in the next round.

Example:
>>> from sympy import symbols, Piecewise
>>> from sympy.printing.pycode import PythonCodePrinter
>>> x, k, c = symbols('x k c')
>>> p = PythonCodePrinter()
>>> piecewise_expr = Piecewise((x, k), (0, True))
>>> pycode_assign_to = p.doprint(piecewise_expr, assign_to=x)
>>> print("Assign to:        ", pycode_assign_to)
>>> print("Expected:   ", "x = (k) if k else (0))")