fix: Move piecewise printing fix to AbstractPythonPrinter
PR #28371 identified and fixed a bug in piecewise printing using the pycode printer, but patched PythonCodePrinter and so left some printers buggy. This commit moves the patch to AbstractPythonPrinter. The regression test has been updated to also test SymPyPrinter, confirming that the whole family of pycode printers are no longer bugged.