integrals: Add a pattern-based integration method
This is intended for cases where SymPy can do most but not all integrals
in a large computation and the user is able to solve the few remaining
ones by hand or can look them up. Then, instead of extracting the
problematic integrals from a large expression and handling them
separately, their solutions can be supplied as patterns and the
integrals will be solved automatically.

The idea is that the user defines exactly the patterns they need and not
that SymPy ships large libraries with predefined patterns. For this, 3
functions (int_patterns, add_int_pattern, clear_int_patterns) are added,
with which the user can interact with the list of patterns.