Replace `if <inequality> is True` with `if <inequality> == True`
The "is True" was added to these at some point to avoid calling
bool(<inequality>), which results in an error if the inequality is symbolic.
However, in SymPy when an inequality does evaluate, it evaluates to the SymPy
true or false objects, not the built-in True or False. So the "is True" check
would never succeed, even when it should.