Comment out failing spde check
This is an example from the book which technically "works" except it uses a
symbolic n and fails the "if n < 0" check (previously this worked on
accident). The example is still useful for double checking the algorithm but
would have to be tested manually. Note that the actual Risch algorithm code
does not call spde() with symbolic values of n.