acos class: Added is_imaginary property check
The is_imaginary property is not implemented for inverse trigonometric
functions such as `acos` and `asin`. Due to this, checks such as:
>>> print(acos(2).is_imaginary)
None
This behaviour can be improved because the real part of acos(x) can be
determined with a simple check on the argument.
$acos(x) \in 𝕀$ if $x > 1$
This has been implemented in the `_eval_is_imaginary` method and the
relevant tests have been added to `test_trigonometric.py`.