Fixed bug in sample for SingleFiniteDistribution
When one chooses the library as SciPy for sampling a finite, discrete
random variable, the current setup forces the elements of the sample set
to be integers. In case the elements are not integers, i.e. the __int__
method cannot be applied to them (e.g. SymPy symbols), we run into an
exception. This is concerning because the Coin RV by default uses SymPy
symbols 'H' and 'T' as default parameters and sampling it without any
other changes leads to this exception. It is also an issue when we are
sampling from a Uniform Discrete RV where not all the elements in the
sample set have the output of __int__ distinct. Some breaking examples
fixed by this are:
>>> import sympy.stats as st
>>> coin_rv = st.Coin('C')
>>> coin_samples = st.sample(coin_rv, size=(N,))  # Wouldn't work before
>>> discrete_int_rv = st.DiscreteUniform('DI', [0, 'a', 1, 1.5, 2])
>>> discrete_int_samples = np.empty(10, dtype=object)
>>> for i in range(10):
...     discrete_int_samples[i] = st.sample(discrete_int_rv)
>>> # Here, we get an accurate sampling of both floats and symbols