dotprint: added switch to use short node names
closes #28580.
sympy.printing.dot.dotprint(...) creates a string representation
of expressions, to be rendered into graph images by other packages,
e.g. graphviz. This requires unique identifiers of the nodes.
These unique identifiers are by default just the expression's srepr,
which grows with the expression. If this string representation is
too long, graphviz raises an error, and the graph is not rendered.
This commit introduces a switch to create short node names of
the form 'node_0', 'node_1', 'node_2', ...

Example:
In [1]: import sympy
In [2]: from sympy.abc import x
In [3]: import graphviz

In [4]: expr = sympy.sin(x)
In [5]: for i in range(8):
........... expr = sympy.Derivative(expr, x) + sympy.sin(expr)

In [6]: string_dotprint = sympy.printing.dot.dotprint(expr,
shortnames=True)
In [7]: graph_source = graphviz.Source(string_dotprint)
In [8]: graph_source.render(filename='mwe_breaking_sympy_graphviz_large', format='svg')

Without 'shortnames=True' in line 6, graphviz would raise an error in
line 8 and not render the graph. The default is 'False', which is the
same behaviour as before this commit.