Consider the case where the result of `factorint` is not prime
When the `limit` option is specified, `factorint` may return a result that is
not fully factored. However, `factor_cache` had been assuming that the result
of `factorint` was always a complete factorization. I updated it so that it no
longer caches results except in places where a prime result is guaranteed.