matrices: fixed matrix summation over null range returning Zero
Fixes summation of matrices with null range returning Zero instead of
ZeroMatrix. For expr.doit(), I added a check for dif == -1. For
expr.simplify(), I added a check for matrices in _eval_is_zero() since
this function is what was returning True previously which led to
expr.simplify() returning zero straightaway. Now, it passes control to
_eval_simplify() which returns ZeroMatrix instead of Zero.