Fix inner Mod square bug (#28744)
Description of issue:
The factors in p are divided between non_mod_l and mod_l with mod_l being the ones that are of type Mod. The reassignment of non_mod_l can include factors of p that are of type Mod meaning that those factors end up being duplicated in both mod_l and non_mod_l. This is what leads to the inner Mod being squared. It only happens if x has the integer assumption because of the is_integer check.

Tests:
Added 2 tests to catch this bug.

Signed-off-by: Harsh Menon <harsh.menon@amd.com>