fix(matrices): handle repeated eigenvalues
Fixes https://github.com/sympy/sympy/issues/28793

The Matrix.charpoly() method splits a matrix into blocks and computes
their eigenvalues separately. In some situations this can mean that
the same eigenvalue is found in two different blocks and they need
to be combined to get the multiplicity correct. This was already handled
but assumed that the expressions for the eigenvalues returned by roots
will always be identical and can fail if an equivalent but not identical
expression is returned like `pi + sqrt(2)*pi` vs `pi*(1 + sqrt(2))`.

If the repeated eigenvalue is not detected then the expressions returned
for the eigenvalues are correct in some sense but not recognising the
multiplicity leads to incorrect results from eigenvects and Jordan
forms.

This commit uses expand() as a simple way to check if two different
root expressions are equivalent. This is not robust enough in general
and so a proper fix would be to have a function (perhaps roots?) that
can find the roots of a list of polynomials correctly accounting for
their multiplicities.

Also adds some type annotations for functions that are called by the
eigenvals routine.