fix(simplify): handle simple integer strings in nsimplify (#28844)
<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that
exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it
is
open. -->

Fixes #23822

#### Brief description of what is fixed or changed

Adds a path to try and parse string inputs to `Rational` before the
integer shortcut in `nsimplify()`. Previously `as_int()` would raise an
Exception for all string inputs, skipping the shortcut, so some unlucky
integers would result in huge, fractional outputs, such as `"4678"`

From the original Issue:

```
In [41]: nsimplify(4678)
Out[41]: 4678

In [42]: nsimplify("4678")
Out[42]: 
                         51      
                         ──      
             7/16  7/40  80 8 ___
13841287201⋅2    ⋅3    ⋅5  ⋅╲╱ 7 
─────────────────────────────────
             17280000  
```

#### Other comments

`Integer()` isn't sufficient as it will round and consume floats!

Two other changes in `nsimplify()` could make sense to review in this or
another PR
* should any given `rational` arg be passed to the main call to
`sympify(expr)`?
* should the `tolerance` arg be used in the internal precision
calculation? (see comments)

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN
and END
statements. The basic format is a bulleted list with the name of the
subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
* Fixed a bug with log of integers. Formerly, `log(-x)` incorrectly gave
`-log(x)`.

* physics.units
* Corrected a semantical error in the conversion between volt and
statvolt which
    reported the volt as being larger than the statvolt.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your
release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->

* simplify
* String integers passed to `nsimplify()` no longer arbitrarily result
in large, fractional outputs due to raising an Exception in the integer
shortcut which wasn't intended for strings

<!-- END RELEASE NOTES -->

---------

Co-authored-by: rfordyce <6064285+rfordyce@users.noreply.github.com>