fix(sets): Remove Set._infimum_key
This method is used to try to sort sets by their infumums but that does
not work in general and is unnecessary anyway since a canonical ordering
is possible without the use of numerical evaluation. This commit removes
the method and adjusts some other code that made faulty assumptions
about the ordering of unions and intersections.